Add simple styling to bup web.
This adds a wrapper set to 960px to the bup web layout. It also
sets widths on the table and columns.

I added a wrapper div to list-directory.html but I think that must
have snuck in with a previous commit.  I am not sure how to fix that
so I will leave it for now.  Sorry!

Adds a wrapper div to the html template.

Signed-off-by: David Roda <davidcroda@gmail.com>