cmd/split: add a --git-ids option.
This lets you provide a list of git object ids on stdin instead of the raw
content.  bup-split then uses a CatPipe to retrieve the objects from git and
hashsplit them.  You could use this as a helper for converting a git repo
that contains a bunch of large files into one that uses bup-style hashsplit
files.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>