Add (sec, ns) timestamps and extended stat, lstat, utime, and lutime.
Replace py_lutimes() with bup_utimensat() and when available, use it
from helpers to provide high resolution lutime() and utime().  Also
provide the utimensat AT_FDCWD and AT_SYMLINK_NOFOLLOW flags from
_helpers.

Add bup_lstat(), and when available, use it from helpers to provide an
extended lstat(), which for now just adds full timestamp resolution.

Change the timestamp encoding to a signed pair of (seconds,
nanoseconds) both internally, and in the archive.