Revert new-style classes
Some classes were changed to "new-style" Python classes in c7a0f06.

Following a discussion on the mailing list about the relevance of such a
change, it was noted that the features that new-style classes bring are
not used in bup, and considering their slightly higher cost in
instantiating them and accessing their attributes, it is decided that we
don't change to using them.

Revert the changed clases back to old-style classes so that all code is
consistent.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>