cmd/save: if file.read() returns an error, don't abort.
Apparently some mis-implemented Linux filesystems (selinuxfs) have regular
files that can be opened for read, but return EINVAL when you try to read
them.  We would throw a fatal exception in that case (since we're not
supposed to have read errors ever, and thus that implies something happened
that we didn't think of) but I guess we'd better make this into a non-fatal
error.  It still makes the exit code nonzero so you can see that something
didn't work, though.

Reported by Zoran Zaric.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>