Revert (most of) "--remote parameter requires a colon"
This reverts (most of) commit c135a5834a9bf9cd9c3382d6489f93e3fdabeafd.

Requiring a colon seems to be too fascist; it makes people think that you
can't use local repositories anymore, which wasn't true: you could just
refer to them as ":/path/to/repo".  But that's just too weird and
non-obvious.  It already resulted in a query on the mailing list, the
avoidance of which is why we added this patch in the first place.  So let's
take it back out.

I kept some minor clarifications and unit test improvements, however.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>