Move commit directories in /.commit/??/
Currently, directories in which we can access files of a particular
commit are placed in each branch directory by which it is reachable.

To avoid possible repetitions of commit directories, move the
directories in a new top level hidden directory named /.commit.

This hidden directory is structured as a two level-deep directory
structure, wherein the first level represents the first byte (two
hexadecimal characters) of commit hashes, and the second level
represents the remainder of the hash.

With this movement, branch directories now contain only symlinks to the
commit directories in /.commit/??/

Also, in BranchList (formerly CommitList), the 'latest' commit was
computed on every iteration over a commit. I moved this calculation up
one level so that it is computed only once.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>