Add (private for now) "bup xstat" command and use it in the metadata tests.
Add a "bup xstat" command that relies on Metadata and prints extended
stat information for the paths provided.  Use "bup xstat" in the
top-level metadata tests so that we don't have to try to accommodate
variations in the local stat command across platforms.

Move the top-level metadata tests from t/test.sh to t/test-meta.sh so
that it's easy to run them independently.