midx: auto-remove midx files that refer to missing .idx files.
Normally an .idx file doesn't ever disappear, but it could happen if you run
'git gc' on your repository.  Which I thought would be a terrible idea, but
apparently it can actually save a lot of space for some people (although it
takes a pretty long time to run).  And when that happens, all your .idx
files move around.  So let's be polite when that happens.  We'll print a
warning the first time, but then shut up after that since the flawed midx
file will just go away.

Reported by Peter Rabbitson.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>