Refuse branch/tag names that start with a dot
In git, branch and tag names are not allowed to start with a dot.

In bup, we also want to enforce this since we want to avoid collision with the
top-level special directories (.commit and .tag).

Also, in save-cmd, there was an unused variable at line 286. 'oldref' is used
and contains the same thing so get rid of 'ref'.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>