Skip over invalid .idx files if we find any.
There's no particular reason to make it fatal; just pretend they're not
there.

Zoran reported a bug where he had (it seems) some zero-length .idx files,
which is weird, but nothing worth aborting a backup over.

Also, fix _mmap_do() to be able to handle mmap'ing a zero-length file
without an error.  It's a trivial and somewhat pointless operation, but it
shouldn't throw an exception.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>