git.py: error when repo's parent dir absent
Currently, when you try to initialize a bup repository inside an
unexistant directory (e.g. BUP_DIR=some_dir/bup_repo, and some_dir does
not exist), bup has to call "git init" to then obtain an error code
which is not very significant to users.

Add a check for the existence of the repository's parent directory and
throw an exception with a more meaningful error message when that
happens.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>