runtests: Apparently $(wildcard) in make doesn't always sort its output.
This meant that on Solaris, tests would be run in a different order, so that
BUP_MAIN_EXTRA (set in tclient.py) wouldn't be set the same as on Linux.

In this case, we know the wildcard will always match something anyway, so we
might as well just let the shell expand it out rather than asking make to do
it.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>