options.py: update docstrings and detail optspec
The docstring on the Options class currently refers to a man page which
does not exist, and still talks about the now-removed 'exe' parameter.
Update this to be more accurate.

Add a docstring to OptDict.

Finally, the options.py file brings a concept of option spec string. Its
construction should be documented. Since we'd like the options.py file
to be a one-file drop-in so that it can be easily used in other
projects, let's document the option specs in the module's docstring.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>