cmd/daemon: pass extra options along to 'bup server'.
Currently 'bup server' doesn't take any options, but that might change
someday.

Also use a '--' to separate the bup mux command from its arguments, so it
doesn't accidentally try to parse them.  This didn't matter before (since
none of the options we were passing along started with a dash) but if the
user provides extra options, it might matter.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>