cmd/mplayer: change cache size from 1024 to 8192.
For some reason, a small number of my .mp3 files don't play with the small
cache size.

It would be better to use something like 'bup fuse' instead of using pipes
to do the caching.  First of all, the named pipe stuff we're doing is
totally gross anyway; secondly, mplayer really prefers to be able to seek
back and forth in the file, and it can't do that in a pipe.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>