cmd/midx, git.py: all else being equal, delete older midxes first.
Previous runs of 'bup midx -f' might have created invalid midx files with
exactly the same length as a newer run.  bup's "prune redundant midx" logic
would quasi-randomly choose one or the other to delete (based on
alphabetical order of filenames, basically) and sometimes that would be the
new one, not the old one, so the 'bup midx -f' results never actually kicked
in.

Now if the file sizes are equal we'll use the mtime as a tie breaker; newer
is better.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>