cmd/date: an initial 'bup date' (bupdate) command.
'bup date' can download a hierarchy of files from an http server, but is
smart enough to download only the chunks it doesn't already have locally.

To do this, you need to run 'bup fidx' to create .fidx files on the server
for any files you might want to download.  'bup date' then looks at the
.fidx files, compares them against local .fidx files to find out which parts
it does/doesn't already have, downloads the missing chunks, and reassembles
the output files.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>