midx/bloom: use progress() and debug1() for non-critical messages
Some messages in these two commands indicate progress but are not
filtered out when the command is not run under a tty. This makes bup
return some unwanted messages when run under cron.

Using progress() and debug1() instead should fix that.

(Changed a few from progress() to debug1() by apenwarr.)

Signed-off-by: Gabriel Filion <lelutin@gmail.com>
Signed-off-by: Avery Pennarun <apenwarr@gmail.com>