Makefile: Fix 'clean' rule
In commit 1df0bdd1ad937, I introduced a problem in the make file: the
chmod operation that gives back some permissions on
lib/bup/t/pybuptest.tmp dies if this directory does not exist.

pybuptest.tmp is only created when running the tests.

when the chmod dies, the clean rule stops, thus not completing the
cleanup, so we must make sure this operation is not fatal if the
directory doesn't exist.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>