bupdate: don't regenerate the local .fidx, just validate blocks we use.
Previously, we would read through the entire source file to generate its old
.fidx, even though we were about to rewrite the whole thing anyway.  Any
blocks that have randomly changed since the last download are probably no
use to us anyway, so regenerating the fidx doesn't really help.  If we find
a block that should match the server but doesn't, just give up on it and
download that block from the server instead.

This avoids the (rather long and slow) indexing phase at the beginning, thus
significantly speeding things up.

As a bonus (assuming you like this behaviour), we won't bother downloading
the file unless it has actually changed on the server since the last
bupdate, even if we've changed it locally since then.  This happens to be
useful in my use case, although I'm not 100% sure it makes sense in
general...