bupdate: change the progbar back to "bytes downloaded" not "bytes written."
The number of bytes downloaded is a bit misleading if there are few changes:
it seems to download at a snail's pace.  That's *mostly* because writing
a 2GB file takes a lot longer than downloading a couple of megs of stuff.

But some basic usability testing indicates that people thought we were
re-downloading the whole image when we weren't.  And that is *much* more
misleading.