tmetadata: the "non existent group name" test didn't make any sense.
There's certainly no reason to expect the file's uid/gid would have changed
after a call that's supposed to fail.  It was passing by pure luck on Linux,
which doesn't have a sticky gid bit causing the newly created file to have
a gid != os.getgid().  But on MacOS, the file was originally created with a
gid != os.getgid(), and so restoring its numeric id restored that, and then
the test failed.

The test is still kind of pointless; it doesn't actually test anything
useful, like (for example) automatic fallback to restoring by numeric gid if
the groupname can't be found.  In fact, looking at the code, it doesn't seem
like it *would* fall back, which is a bug.  But I'm not going to fix that
right now.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>