tgit.py: provoke ENOTDIR rather than EACCES in test_check_repo_or_die().
Replace the objects/pack directory with an empty file to provoke an
ENOTDIR error from stat('objects/pack/.').

Previously the code changed the permissions of the test directory to
0000 in order to provoke an error other than ENOENT (i.e. EACCES), but
that doesn't work when the tests are run as root or fakeroot.

(As Gabriel Filion pointed out, the chmod of the testdir is no
longer necessary, so I removed it and squashed that into this patch.
-- apenwarr)

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Signed-off-by: Avery Pennarun <apenwarr@gmail.com>