Earlier "negative timestamp" patch had a 64-bit timestamp in the test.
The date in the comment is correct - for -0x80000000.  Sadly, the *code*
actually said -0x90000000.  That works on 64-bit systems (and filesystems)
not not 32-bit ones, where python gives an encoding error.

In any case, based on the comment (June 10, 1893) it seems tat -0x80000000
must have been the intended value anyway.  Now 'make test' passes on 32-bit
Linux again.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>