ftp: don't output trailing line for 'ls'
'ls' is currently the only 'ftp' subcommand that outputs a trailing
newline before the prompt is re-displayed. This is cause by the use of
"print" to output a string that already contains an ending newline.

For a matter of consistency of output, make 'ls' output without that
extra trailing newline.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>