save-cmd.py: remove redundant _push()
Remove a duplicate _push() from the "first_root" code.  This could cause
the creation of a tree that was immediately _pop()ped by the "finish the
current sub-tree" code, and then reintroduced (as a duplicate parent
entry) by the "start a new sub-tree" code.

Instead, just wait for the "start a new sub-tree" code to call the
relevant _push().

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>