subtree-hash: handle non-ASCII path chars
Pass "-z" to git ls-files to turn off git's core.quotepath, so we can
handle non-ASCII path characters directly, and rewrite the code in
python to make that easier.

Thanks to Jiří Martínek for reporting the issue -- that
test-redundant-saves.sh would fail when there was a "ý" character in the
source tree path.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>