Create t/sampledata/var/ and version it
Maintain a new t/sampledata/var/ (via t/configure-sampledata) that
contains any test data that we don't want to or can't commit to
git (i.e. symlinks, and other dynamically generated data).  Move all of
the existing generated data there, and delete var/ entirely on clean.

Control the creation of var/ with make, via the existence of
t/sampledata/var/rev/vN.  Whenever we change the content, we'll change
N (currently 0), which will force the directory to be recreated.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>