Revert "Avoid fadvise (since it doesn't work as expected)"
On some systems (the reported system had 2GB), not forcing out all of
the data traversed during save dramatically slows down save operations,
possibly due to competition with access to the indexes, etc.  So restore
the use of fadvise_done() for now.

This reverts commit db68c4fd4dc3087eb3f0b6e5846629cbf02d0b27.

Thanks to Tilo Schwarz for reporting the problem.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>