test.sh: separate index and split/join tests
Move the index tests to test-index.sh and the split/join tests to
test-split-join.sh in order to increase the potential parallelism (and
modularity/isolation).

Here, this brings the "make -j check" time down quite a bit.  Before
test.sh was always the last thing running, and the index tests took most
of the time.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>