thashsplit: don't assume MINCORE_INCORE is defined
We had intended to guard the use of MINCORE_INCORE with tests for the
presence of the underlying function call, but we missed the hashsplit
tests.

Fix it by allowing the tests to pass in the mask value they already
depend on (i.e. 1).

Thanks to Michael March for reporting the problem, Patrick Rouleau for
confirming it, and Markus for proposing an alternate solution.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>