Choose 2M when SC_ARG_MAX is -1 (unspecified)
When sysconf() returns -1 for a limit, that means the limit is
"unspecified".  In that case, choose 2M for SC_ARG_MAX instead of the
POSIX minimum, matching (somewhat arbitrarily) the value on a current
16GB Debian amd64 system.

The POSIX minimum was unnecessary, likely inefficient, and causing
unnecessary failures.

Thanks to Mark J Hewitt <m.hewitt@computer.org> for reporting the
problem and tracking down the cause.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>