Generate bup-python from Makefile not configure
Don't generate cmd/bup-python in configure.  Instead, just define a
bup_python config variable (via config/config.vars.in), and use it to
generate bup-python from ./Makefile.

In support of that, include config/config.vars in ./Makefile, and add
a rule to generate it.  This simplifies some of the dependencies, and
removes a level of indirection when determining the correct python to
use (i.e. PYTHON via $(shell)).

This does mean that a "make clean" on a clean tree will run configure.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>